requirements.txt
# 农村女性就业市场动态演化机制研究项目依赖
# 基于平均场博弈与主体建模思想
# Python 3.11.0

# 核心数据科学库
numpy==2.3.3
pandas==2.3.2
scipy==1.16.2

# 数值加速（用于MFG和KFE求解器的JIT编译）
numba==0.62.0

# 统计分析（用于匹配函数的Logistic回归）
statsmodels==0.14.5

# 可视化
matplotlib==3.10.6
seaborn==0.13.2

# 配置文件处理
pyyaml==6.0.3

# Copula建模（用于虚拟人口生成）
copulas==0.12.3

# 多进程增强（支持闭包序列化，用于DE并行校准）
pathos==0.3.3

# 测试框架（可选）
pytest==8.4.2
